[SemanticMetaTags] Add dependance on SemanticMediaWiki
Doint need to use composer just add dependance on this extension.

Currently there is a problem with doing it in other test other then this
one.

So untill we fix the way we install the thing s first smenaticmediawiki
will fail so might as well go with the right test that works and passes
the database.

Change-Id: Id7b697a7fcc87970416988c8a3bdc9dc1a77e711