Update patch set 1
Patch Set 1:

Sounds wise and can be generalized to the npm job as well.

A word of caution though, if we were to add composer test jobs to all repositories, and when they get migrated to Nodepool disposable instances, that mean we will consume instances essentially to just do "exit 0".

It is not ideal, but I guess it is up to us to have composer.json test entry point added to all repositories.

Patch-set: 1