Make "recheck" magic word less restrictive for Code-Review votes
Currently, if you want to retrigger jenkins and want to change the Code-
Review vote, you would need to first change the Code-Review vote in gerrit
and then add a new comment with the "recheck" magic word. That isn't really
cool :P

This change changes the underlying regex to work with any change of the Code-
Review vote (+2, +1, 0, -1, -2) and a recheck as a comment, so you can recheck
and change the Code-Review vote in a single comment.

The regex can be tested at https://regex101.com/r/gG1cK8/1

This would, with a little change, work with the "Verfied" flag, too, but I
can't think about a use case, where a reviewer want to set the Verified flag
and also trigger a new run of Jenkins for this change, so this is not subject
of this change.

Change-Id: I2e0ea1069c40bc539fe848abf0919b602613f746