Update patch set 1
Patch Set 1:

Ok, I'll look into it, how to write tests (tests are always good :)), and if I've questions I ping you in irc :P

Yes, a +2 and recheck as a comment would trigger 2 test runs in zuul, one for the recheck (in the check queue) and one for the gate-and-submit pipeline. But, in my point of view, there are valid cases, where a recheck and a +2 makes sense (e.g. if a change has an unmerged dependency: The change wouldn't be mergeable, so the gate-and-submit pipeline wouldn't be started (iirc) and maybe I want to add this +2 and want zuul to run the test pipeline nonetheless). So, I think we can trust the reviewer, that he uses the +2 + recheck only in valid cases, otherwise the only "problem" is, that zuul would run the test suite twice (which isn't really cool, but not anything which will result in a disaster :P. And if it is mis-used more and more, we can easily remove the trigger for a +2 ;)

Patch-set: 1
Label: Verified=0