Update patch set 1
Abandoned

Being redone by Kunal. He already sent a patch to use a single parameter-function which was the reason I reverted the 5.5 patch.

Note for later: instead of having "Revert, "Revert foo"", reuse the original commit message and tweak it :-}  That is easier to understand what the change is actually going to do.

Patch-set: 1
Status: abandoned