Update patch set 1
Patch Set 1:

Ok if we add assert.expect manually to all the test files in ValueViews extension wikidata it should pass. It worked for me I just copied that patch over too here to test and it failed at it could not identify one of the dependacys but that isent wikidata or ValueViews because I didn't copy it all over just wanted a quick test.

Patch-set: 1