Update patch set 6
Patch Set 6:

Yeah that might be solving an issue that is gone on with install.php failing because the database is not all setup yet but hooks are kicking when editing the Main_Page

The order needs to be checked,  it is done differently on other jobs iirc.

We would want to look at all the jobs and see in which order steps are executed. Settle on one and make sure all jobs use it.

The original orderwas probably to install mediawiki/core first then load extensions and run update.php to setup the rest of the db.

Patch-set: 6
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0