Update patch set 7
Patch Set 7:

The call to wfLoadExtensions is only added to LocalSettings.php during mw-apply-settings. Maybe some of this is because the LodalSettings.php from a previous run is still around and the installer doesn't overwrite it?

I remember there was a bug in a few extensions, that they used composer to load themselves when the composer autoloader was included. Extensions should not do that, that leads to problems when the composer-merge-plugin is used. Wikibase was one of them and I think some SMW extensions also did this. Please check if this is the case first.

Patch-set: 7