Remove mediawiki-core-jslint job
It was kept as convienience but is no longer useful because the
JSHint version used by mediawiki-core npm-test is no longer compatible
with the legacy global version pre-installed in the CI environment.

That install was already deprecated, but we kept it for now as
a convenience, but it's now failing on every commit in the 'check'
pipeline and no longer serving any purpose.

It'll effectively come back for the 'check' pipeline when the
'test'-pipeline jobs are migrated to the secure Nodepool environment.

Change-Id: I64cab45ff3b678072a78e0b1a8bdd10cca4deb52