Update patch set 2
Patch Set 2:

In a follow up patch if everything passes we would need to split the php55 tests for wikibase into a separate template so we can blacklist it from running on branches REL1_23 - 1_26. We also need to do php53 to only run on those branches but not on master.

Patch-set: 2
Label: Verified=0