Update patch set 2
Patch Set 2:

I think we are better sticking with the find | xargs php -l for now.  This patch looks fine but it has too much overhead:

* clone/refresh vendor
* run composer
* parallel-lint which runs on ALL php files.

The mediawiki-core-php55lint we run takes only a few seconds:

* git clone (workspace being kept between run)
* fetch / checkout
* look at php files being changed in HEAD
* only php -l the changed files.

It has some slowness though whenever a workspace need to be populated via a full clone, but it is usually quite fast https://integration.wikimedia.org/ci/job/mediawiki-core-php55lint/buildTimeTrend


So we will want to keep that in mind, whenever we start running 'composer test' for mediawiki/core we will want to only have it run on files changed in HEAD or get it out of the 'test' entry point.

phpcs should be optimized as well to only deal with files changed in HEAD. That is another topic though.

Patch-set: 2
Label: Verified=0