dib: provision openJDK
The Nodepool image lacks a JRE installation preventing jobs depending
on java to run on it.
Jenkins also depends on a JRE although it is able to provision it by
itself when the slave connection is established over SSH.

Install the JRE package and also include the Jenkins requisite which
provides the JRE as well.

Bug: T126246
Change-Id: I520238574777f712ec9b2dd5d2cb8438244ef930