Update patch set 1
Patch Set 1:

Eek! Another reason for ignoring the exit code is that some tests fail when run with xdebug/coverage.  On failure nothing is published.

But yeah wiping the coverage is terrible. So we should indeed stop ignoring the failure, identify failing tests and fix them then find a way to report somewhere that the job is broken.

Patch-set: 1