Update patch set 1
Patch Set 1:

Were in the process of removing jslint tests, But first we need to replace all tests that depend on this with alternatives. We are going to use npm here to use jshint and jsonlint that way in future you can add tests as you like through npm. Please note non whitelisted users will not be able to run jshint and jsonlint once npm is switched on this repo.

Npm will soon be allowed to be run by non whitelisted users.

Patch-set: 1
Label: Verified=0