Update patch set 1
Patch Set 1: Code-Review-2

This doesn't actually make a difference for what you describe in the commit message. This only affects the order in what repositories are downloaded, not the order in which they are loaded.

But more importantly, extensions should be able to be loaded in any order, and check dependencies once everything is loaded. If that's not the case, they should be fixed.

Patch-set: 1
Reviewer: Legoktm <458@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2