Update patch set 1
Patch Set 1: Code-Review-1

phpize is a command shipped by the php Debian package. On Precise that is Zend 5.3 and on Trusty it is currently Zend 5.5.

I have no clue how extensions are build for HHVM, but maybe the template should be renamed php55 build, changed to Trusty and invoke /usr/bin/phpize5

Later on we might want a slave script phpize on a similar model as the php one that vary based on PHP_BIN. So we can potentially build an extension against different versions of PHP.

Patch-set: 1
Label: Code-Review=-1