Update patch set 1
Patch Set 1:

(2 comments)

The logic is wrong see inline.  Anyway the order should not matter and extensions should really be loadable in any order.

Ideally, I would even randomize the order of the file so the tests randomly fail. But that might be too evil.

I agree with Kunal that extensions failing with the current order should be fixed.

Patch-set: 1