Add npm-old template and builder
Reason im doing this is so that repos such as OOJS use composer in there
npm test but composer is not supported on nodepool so as a workaround
tests that carn't be migrated to nodepool should use this test and the
tests which can use noodepool then should use npm-node-4.3 too.

But they will need to update gruntfile.js for them to update the tests.

This will help alot in the migration to npm 4.3.

Change-Id: Id29e17e76f49ef8048ca9196f0f9361a77671d20