[mediawiki/event-schemas] Dummy definition
Whenever a change has a Depends-On: on a repository which is not known
to Zuul layout, the change item is not properly populated and miss the
project name.  That cascades in a deathloop with the event processor
looping over and over and raising an exception about the project name
missing.

Workaround it by definining a dummy mediawiki/event-schemas project with
solely noop job.

Bug: T128569
Change-Id: I7698df50ecdc94430331cd4d85b3954361813449