Update patch set 2
Patch Set 2:

Looking at the Zuul layout diff, that change the regex which was:

 mediawiki-extensions-php55> (?!REL1_2[3-6]$)

To:

 mediawiki-extensions-php55> (?!REL1_23|REL1_24|fundraising/REL.*)

So that will bring cause that php55 job to run on REL1_25 and REL1_26.   Need to narrow the scope. I supposed there is a rule for *php55* which is thus overridden.

Maybe the easiest is to do a specific rule for:

 ^mediawiki-extensions-php55$ 
 branch: (?!REL1_23|REL1_24|REL1_25|REL1_26|fundraising/REL.*)

Patch-set: 2