Update patch set 1
Patch Set 1:

Depends. We might be able to do a proper git-clone without exclusion of submodules. Depends on whether nodepool has a git cache. I think it does. So the reason for excluding submodules (which was also the reason we had a separate job for VE which needs submodules) might be moot now.

Patch-set: 1