Update patch set 1
Patch Set 1: Code-Review-1

If it is at all feasible to fix this without this change then that is preferred. It is not acceptable to have to add no-ops for every repo created. We never did this and we should not start now. Even if we'd do it now, we'd still have a deadlocked Zuul everytime a repo is created requiring a manual change, which doesn't scale.

This needs to be adequately addressed in Zuul - T128569.

It seems that a restart helps sometimes. Albeit losing the active queue.

Patch-set: 1