Update patch set 1
Patch Set 1: Code-Review-1

@Hashar we could do what @legoktm did and run the test on all repos. 

But before we do migrate even if all the repos pass which I think all do we need to put more resources to Jessie on the day we switch before merging this to make sure that there is a smooth transition. We doint want the https://integration.wikimedia.org/zuul/ to be backedup with a ton of jobs taking hours to complete whereas it would normaly take a minute or two to run.

Patch-set: 1
Reviewer: Paladox <1665@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1