Whitelist user Robert Vogel
He is a long term user See
https://gerrit.wikimedia.org/r/#/q/owner:vogel%2540hallowelt.biz+status:merged,n,z

He is also part of this group

https://gerrit.wikimedia.org/r/#/admin/groups/497,members

Plus he frequently uploads patches See

https://gerrit.wikimedia.org/r/#/q/owner:vogel%2540hallowelt.biz+status:open,n,z

Also it would be useful for him or her for reasons it can prevent the
extension from failing. It can also help fix the failing unit tests in
check experimental so we can switch the repo to unit tests.

Change-Id: I1eea7b3c7ec4a3601bbc257bee54a2a481d38ed4