Allow decoupling of npm-node-4.3
Zuul puts projects in the same merge queue whenever they are sharing a
job. In lot of cases that is really unwanted.

A challenge was to find a way to avoid repeating ourselves between the
generic and variadic job templates.  I did an experment using YAML
merging on https://gerrit.wikimedia.org/r/#/c/243164/ which seems
successful.

Mark 'npm-node-4.3' with a YAML alias.
Create job template '{name}-npm-node-4.3' that merges in the job.
Override the 'name' variable in the template definition.
Reapply 'triggers: - zuul' since JJB strips it in this corner case
usage.

This trick avoid duplication and let us define more specific jobs. Do so
for the wikimedia/portals repository which only triggers npm job and
asked to be decoupled (T129591)

Expand the Zuul project template. We cant pass the full project name
yet, will be possible later with Zuul patch:
  https://review.openstack.org/295237
  Provide 'longname' to project templates

Bug: T107529
Bug: T129591
Change-Id: I66b2a86f506bef757fa9a90ee51f9746a474a79d