Update patch set 1
Patch Set 1:

legoktm : mind looking over my shoulders for this patch? It is basically rewriting the way we find unsafe jobs.  Previously each iteration of the loops invoked helper methods that also did loop over the whole workflow to retrieve a single item.  That was a lot of unnecessary lookups.

The rewrite does a single request to get a map of "project -> pipeline -> jobs" then just inspect that dict of dict of list.  That is tremendously faster :-}

Patch-set: 1
Label: Verified=0