Update patch set 1
Patch Set 1:

Reommended by hashar for repos that doint have js files.

The reason is because once the migration to npm is complete we will remove all jshint and jsonlint tests from ci. Since the idea is for npm to be open to everyone including non whitelisted users, but currently that is not possible yet. But should soon be.

Patch-set: 1