Update patch set 4
Patch Set 4:

JanZerebecki: this was brought to my attention on today's scrum of scrums. I created a task so that further work can be tracked and to be sure that this isn't forgotten. Can we merge this as-is and then follow up on https://phabricator.wikimedia.org/T131264 regarding the use of composer-merge-plugin? If there is no significant reason to block this change then I'm happy to merge it

Patch-set: 4
Label: Verified=0