Update npm to 2.x
Currently it is running npm 1.x on nodepool, it should run npm 2.x so this
is a temp solution until someones builds npm 2.x into the image or someone
uses nvm.

We need 2.x since it would be a downgrade using 1.x not 2.x. 1.x is also
no longer supported and a security risk to use.

Change-Id: I376463c0e9bc482bb18619ee64b8c977c37e8328