dib: provision npm 2.x
Jessie provides npm 1.x but we had NodeJS 4.x backported which is
normally shipped with npm 2.x.  Additionally the version provided by
Debian is known to be crippled with bugs for any non trivial case and is
unusable for a few of our repositories.

It is quite unlikely npm 2.x will be packaged for Debian any time soon.

In the DIB image, provision npm from NodeJS the same way we do it on
permanent slaves.

Note: needs I391bc4f5f42bed2f1dd76f0e1efc337dbed97bd7 which removes
grunt-cli, it is unwanted on Nodepool instances and is not needed (repos
should list it in their devDependencies).

Bug: T124474
Change-Id: I662d180e2d5ac4664d764b94df493998e3379c60
Depends-On: I391bc4f5f42bed2f1dd76f0e1efc337dbed97bd7