Update patch set 1
Patch Set 1:

No you doint have to use it. But it is already there and it would make updating tests harder if we decide to update the tox-jessie template.

check:

can only run basic tests like jsonlint and jshint and phplint and some other basic tests

Whereas tox-jessie cannot be run for non-whitelisted users in check:

But hashar is soon going to open up npm to non-whitelisted users soon.

Patch-set: 1