Update patch set 2
Patch Set 2:

Well I would rather not block https://gerrit.wikimedia.org/r/#/c/279367/ on potentially maybe speeding up Scribunto soon and then maybe adding it to the job.

It should be removed until it can be added without causing issues.
Or just add the dep and deal with the speed hit.

Also the patch in question is introducing one of the core mediawiki tests that you are talking about. Thats how we know ZeroPortal is badly configured (missing the dep)

Patch-set: 2