Update patch set 1
Patch Set 1:

Will this run the unit test of the extension? If yes, please don't merge this change yet, there is one broken unit test in the extension which I don't know how to fix because it's caused by an insert in the setup of MediaWikiTestCase (see the TODO in `CognateStoreTest::setUp` method).

Patch-set: 1
Label: Code-Review=0