Set up CI for event-schemas repository.
We've set up some basic tests for event-schemas repository,
so now need to set up a Jenkins job. The difference with normal
npm testing is that event-schemas are not a real npm package, so
we've put a package.json inside the test directory. The job needs
to do the same thing as other npm testing job, but before it has to
change directory to the 'test' dir inside the WORKSPACE.

Bug: T124438
Change-Id: I05b46d48a05bc1a3256099be76f5acfc06171b9e