dib: composer and HHVM on Jessie
Get composer installed on Jessie by moving the include out of condition.

We should install php5 using mediawiki::packages::php5 but that one
ships php5-fss which is not needed on Zend 5.5+ and is not built on
Jessie (see T95002 for discussion). Thus install HHVM to provide a PHP
interpreter.

Bug: T128092
Change-Id: I41f7aaee6496a59ae8662d9e388955b108b7dbb0