Migrate parsoidsvc-php-parsertests to Nodepool
Move it to Jessie Nodepool instances with HHVM and NodeJS 4.3.

The parser test 'Running test Image with page parameter' fails due to
lack of pnmtojpeg command. It is a recommendation for imagemagick and in
prod is installed via mediawiki::packages::multimedia. Add the package
to the dib images.

Delete: parsoidsvc-php-parsertests
Create: parsoidsvc-hhvm-parsertests-jessie

Bug: T126992
Change-Id: Idd5fb9e1485c7d838ef5dcba6aa747b007b52557