Make mwext-qunit to timeout earlier and not fail
There is some nasty issue going on with the mwext-qunit which sometime
idles waiting to close chromium sub processes.  The 30 minutes is to
long and there is no point in failling the build in this specific case.

Lower timeout to 7 minutes.
Do not fail the build on timeout.

The later should be removed when the root cause is determined, since a
timeout should really be noticed.

Bug: T119143
Change-Id: Ic32c9aaa651557984338e64c5ea70b752cf41f15