Update patch set 7
Patch Set 7:

Since we should be migrating to Jessie. trusty is on nodepool to allow us to migrate tests to trusty but on nodepool that we carn't yet do in Jessie.

Plus this test is still on Ubuntu trusty not trusty nodepool.

Patch-set: 7
Reviewer: Paladox <1665@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0