Enable zuul/jenkins to run unit tests in skin projects
This patch will allow us to test skins like we do with extensions. But it
will test differently to extensions due to the fact skins are different to
extensions.

Skins can also depend on other skins to test against same as extensions
can.

Ive also added mediawiki-skins-(qunit|phpflavour) to gate and submit.

It is experimental so i have not added it as a test in mw core just in
experimental:

Metrolook, MonoBook and Vector are in gate and submit for testing but will
only run if you do check experimental

Depends-On: Iac5b1be23dd9b657f43bd7d5a3fe5ddff326e25f

Bug: T68926
Change-Id: I1be4454b1aa7935ec2e9ac3703eee3907e5101e8