mwext-mw-selenium jobs on Nodepool
Previously sent as d9b9e86.  This time introduce the Nodepool jobs in
addition of the one running on permanent slaves. They are only triggered
in experimental pipeline or made non voting.

Add a JJB macro mw-selenium-nodepool which is an adaptation of
mw-selenium macro, though lighter and appropriate when running in fresh
env, namely:
- drop GEM_HOME and GEM_PATH mangling
- use bundler from Jessie instead of installing it from gem
- copy pasted the cucumber line. Should be replaced by 'rake selenium'

mediawiki-selenium-integration-jessie is added as a non voting job.

Create:
    mediawiki-selenium-integration-jessie
    mwext-mw-selenium-composer-jessie
    mwext-mw-selenium-jessie

Bug: T137112
Change-Id: I699d5be1ad3d57637f49add0dbfbfd94a23ff536