nodepool: last distupgrade should keep config files
We are doing a late dist-upgrade to make sure we have the latest
package, but it will choke whenever one of our puppet recipe previously
altered a config file which the package wants to upgrade. The snapshot
ends up being stall with:

 Setting up foobar (1.0.0+deb2)
 Installing new version of config file /etc/foobar/default.conf

 Configuration file '/etc/foobar/default.conf'
  ==> Modified (by you or by a script) since installation.
  ==> Package distributor has shipped an updated version.
    What would you like to do about it ?  Your options are:
     Y or I  : install the package maintainer's version
     N or O  : keep your currently-installed version
       D     : show the differences between the versions
       Z     : start a shell to examine the situation
  The default action is to keep your current version.

That prevents Nodepool from completing the snapshot.

Instruct dpkg to keep the old configuration if it got altered by passing
the option DPkg::Options::=--force-confold . That is how Puppet fix it.

Change-Id: I4503f5925ec00e7a779c9bc1232f6bb351fc474a