Update patch set 5
Patch Set 5:

This change replace the old jobs entirely. The new ones are not known to pass on Nodepool instances and there is no straightforward way to test them.

So maybe we should keep the known to work mwext-mw-selenium job  and repurpose the change to add the new job in parallel and add them to the experimental pipeline. Will let us easily trigger them and iterate as needed.

Patch-set: 5