zuul: run noop on gate for mw-checks template
The gate-and-submit for mw-checks template is solely made of jobs that
have file filters.  If one send a patch that does not touch any files
matching those filters, no job get triggered and the change is never
merged.

Add noop to mw-checks gate-and-submit so we have a guarantee that at
least one job is triggered.

Change-Id: I0c761f3dcfcb89d6b1f705537939bcd9c5102dfb