Switch apps-android-wikipedia jobs to Java 8
openjdk-8-jre is available on the permanent Jessie slaves. Jenkins
global configuration has been updated to refer to it:

Name       ->  Debian - OpenJdk 8
JAVA_HOME  ->  /usr/lib/jvm/java-8-openjdk-amd64

Update the three related jobs to run on Jessie (DebianJessie label) with
JDK set to OpenJdk 8:

    apps-android-wikipedia-lint
    apps-android-wikipedia-publish
    apps-android-wikipedia-test

The apps-android-wikipedia-periodic-test job runs on label
AndroidEmulator which is solely on integration-slave-trusty-1023. Will
be migrated later on, we need a dedicated instance running Jessie to
migrate it to.

Bug: T138506
Change-Id: I4b0c34bb108a15e85ccf0ae08a0bbe31e89a2b57