Build step to install Android SDK
The android-emulator has a build step to automagically install the
Android SDK on a slave when it is missing. There is at least one job
using the android-emulator which thus take care of installing it but it
is not necessarly running on all slaves.  When a job depends on Android
SDK it might miss it.

Since JJB does not support the build step, inject the raw XML to the few
jobs that rely on Android SDK.

Bug: T138506
Change-Id: I1b9b2430d571fee55fddd424fedbdaae92c86719