Migrate coverage reports to doc.wikimedia.org
Move the coverage reports from integration.wikimedia.org to
doc.wikimedia.org so they can share the same backend in the Varnish misc
cache. The integration.wikimedia.org domain is going to be moved to
another host.

Change the rsync commands to point to a different docroot:

    - org/wikimedia/integration
    + org/wikimedia/doc

When we get a new web host, we will change the publish-on-gallium job.

Update jobs:

    mediawiki-core-code-coverage
    oojs-core-coverage
    oojs-ui-coverage
    phpunit-coverage-publish
    unicodejs-coverage
    visualeditor-coverage

Bug: T139620
Change-Id: I8df153a62101a8dc7f418bab894dd9f5a7223da2