Fix zuul being pinned to a specific commit
Use Zuul from the branch patch-queue/debian/precise-wikimedia which has
the appropriate patches applied.

Replace concept of triggers with connections.

The FakeConnection class would use as driver the name of the connection
for sake of simplicitly.

Registered the fake connection with the scheduler helper method
registerConnnections

Since the layout validation now requires the Zuul config file to have a
proper Gerrit section:
* add a dummy zuul.conf under tests/fixtures
* update integration-zuul-layoutvalidation to point to it

Bug: T136610
Signed-off-by: Paladox <thomasmulhall410@yahoo.com>
Change-Id: I5f51e6026cfa08566c023b3fdcfbdfe3965f4a91