Update patch set 20
Patch Set 20: Code-Review+1

I have tested this and it works.

If you haven't done +2 for code review, then doing recheck gate will not merge it or start it.

So unless someone +2 and it failed to merge for some other reason then a whitelisted user can recheck gate.

Tested on

http://gerrit-test.wmflabs.org/gerrit/#/c/13/

and

http://gerrit-test.wmflabs.org/gerrit/#/c/14/


This is now ready to be merged.

:) :)

Patch-set: 20
Label: Code-Review=+1