Update patch set 4
Patch Set 4:

That is large of overhaul of the jobs building the debian packages. I have sprinted a working 'debian-glue' job over the last few days/week-end and this change merely drop all the legacy crap.

That change all currently configured repo to trigger the 'debian-glue' non voting job and have it triggered in the 'test' pipeline.

Patch-set: 4