debian-glue: lintian now fails instead of UNSTABLE
Switch the lintian publisher from junit to xunit. The later would fail
the build whenever a test fail regardless of the status of the previous
build (which is usually from a random patchset/different repo).

That should make debian-glue to mark the build as FAILURE whenever one
lintian issue is encountered.

Change-Id: I3202fbd5185b04f19f92c688937befd749fb7f18