debian-glue: Merge 2 shell builders, set distribution correctly
generate-git-snapshot would muck up the workspace by checking out
master, making the call of dpkg-parsechangelog parse an old changelog
resulting sometimes in incorrent distribution setting. Move the call of
dpkg-parsechangelog before the generate-git-snapshot call to avoid that.
In order for that to succeed we need to merge the 2 shell builders
containing those 2 calls. This should be ok given the first one is a
very small builder

Change-Id: Ic3180426ee758803914f53a0891b94120123f011