debian-glue:
zuul-cloner checkout the patches in a detached branch not in a matching
local branch.

generate-git-snapshot at the end of its execution, checkout $GIT_BRANCH.

That causes the workspace to be set to the tip of the branch instead of
the patchset we want to test.  It causes a lot of troubles such as
parsing debian/changelog from tip of branch instead of the patch.

Set $GIT_BRANCH=$GIT_COMMIT so the patchset is checked out.

Bug: T141607
Change-Id: Ief16420d46ab47832d332aac516a687dbf31d996