Update patch set 1
Patch Set 1:

(1 comment)

Just overwrite the old names based on trusty.  But yeah looks like that is all what needs to be done.

For completion, we will have to puppetize the hack in /etc/hosts to have localhost to always resolve to the Ipv4 127.0.0.1.  But maybe something can be fixed in the Jenkins plugin as well.

So reuse the name, deploy with JJB and you can manually retrigger it to see the result. If all fine please self merge !

Patch-set: 1
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0