debian-glue: let us override the build timeout
Add the parameter BUILD_TIMEOUT defaulting to '30' (minutes) to the
debian-glue jobs and pass that to the build timeout plugin.

Have Zuul inject a value of '180' (minutes) to debian-glue jobs being
built for operations/debs/contenttranslation/giella-sme

Update:
  debian-glue
  debian-glue-non-voting

Glitch: the Zuul parameters are added by triggers: zuul are added back
somehow.  It is a non issue though.

Bug: T143546
Change-Id: I272eee0f5a859a80673dea02ba6507645e56ce3f