Always set BUILD_TIMEOUT for debian-glue jobs
Although the jenkins jobs have the parameter with a default value, when
it is triggered by Zuul/Gearman, the defaults of the jobs are not merged
in.  The BUILD_TIMEOUT has been set as a global variable to ensure it is
always set.

Set it from the Zuul parameter function for double protection.
Add a comment in the job definition.

Bug: T144094
Change-Id: I815f66bda15a3ffaed61b190a2c1a2b0921a5fe6