rake: tweak files filter
54b4a024 was a bit too strict for the rake job files filtering.

Split the oneliner regex to a list of entries to match.
Note that re.match() is being used (match at beginning of string)
Whitelist 'rakefile', 'Rakefile', and '/spec/', anywhere in the path.

Add a test covering the filter. It collects assertion errors to report
them all at end of test.  Makes it easy to see all cases.

Bug: T144325
Change-Id: Ibdfa1c8f56bbb1e609ee479dde5d997ed53bfe6f