Update patch set 7
Patch Set 7: Code-Review-1

This is going to have unwanted side effects.  Whenever a file from the polyfill is changed, the job is going to be skip entirely and thus not linting other php files it might have introduced at the same time.

The php55lint job in Jenkins is very lame and does not support any exclusion list.   If that is needed, we should move to use the 'composer test' entry point and jakub-onderka/php-parallel-lint which has support for exclusions.

Patch-set: 7
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1