Puppet doc now ignore /bin files
The puppet doc command pass all files to the command line but rdoc
eventually stops when processing a file under /bin with:

  (Errno::ENOENT) No such file or directory - dummy.rb

Upstream bug:
https://tickets.puppetlabs.com/browse/PUP-3261

Use RDOCOPT to inject an extra --exclude that would prevents rdoc from
processing the modules bin directories.

While at it pass to puppet --trace to get some kind of clue as to why it
sometime fails to process a manifest.

Bug: T143233
Change-Id: I81bc534d6564c8ba6363479ae1b95e74c279a316