ZeroPortal/Banner stop using specific jobs
Both ZeroPortal and ZeroBanner are in the shared jobs defined by
'mediawiki-gate'.  It has a bunch of extensions which cover the list of
dependencies expressed in the Zuul parameter function.

Remove generic unittests for both repo.
Drop the list of dependencies for ZeroPortal, no more needed.

TODO: Drop ZeroBanner list of dependencies as well but that is used by
MobileFrontend itself used by others.

Bug: T145227
Change-Id: I31b83af88fd78d71c5fef1a853fd7180a4d43d40