mediawiki: Merge parsertests job back into main phpunit job
This was originally separated because the parser tests were too slow.
Since recent refactoring they are much much faster. By accident, the
refactor also broke --exclude-group=ParserTests (fixed in I8131c3b13).
This meant that for the past week we've already been running parser tests
in the main phpunit jobs as well. And still, it wasn't the slowest job.

Change-Id: Ic2934746be68c06c1fc747e93f4fa73200eba652