[MobileFrontend] exp job for npm run lint:modules
Joakino is working on an experimental job that lint the JS sources
dependency graph against ResourceModules.  The job requires a clone of
MediaWiki core so we introduce yet another job that deals with it.

Trigger the job manually via experimental pipeline.
Set a branch filter to NOT run it on REL1_27 and below and not run it on
wmf branches.

Creates:
   mwext-MobileFrontend-npm-run-lint-modules

Bug: T146748
Change-Id: I1c8e1562f59b5a8554e6b06f4ad55f25fade9ece