Update patch set 7
Patch Set 7: -Code-Review

Second thought!

The mwext-testextension-hhvm cant pass since it will use mediawiki/vendor@master which miss the dependencies.  So it kind of make sense to use composer in this specific case. I am not sure how useful it is though.

The job mwext-donationinterfacecore-REL1_27-testextension-zend55 covers that the DonationInterface@master t is guaranteed 27to work with mediawiki/vendor@fundraising/REL1_27.

The mediawiki-extensions-hhvm has little purpose here. It is meant to test together a few main stream extensions used on the wikis.  I dont think it is appropriate for fundraising.


Maybe we can write down somewhere what we want to be tested, have a full list of all the dependencies and create a new job with the proper branch hints.

Patch-set: 7