dib: craft our own hiera lookup data
When building an image, we dont rely on external datasources such as
wikitech or httpyaml on the labs-puppetmaster.  The first is never used,
and the second is unreacheable from the internet.

Craft our own custom little hiera configuration which solely rely on
puppet.git /hieradata/

That fix build on Jessie.

Change-Id: I2baaa225521c5648f957d3e9678f62b45ee8011d