Update patch set 2
Patch Set 2:

I did a recheck on https://gerrit.wikimedia.org/r/#/c/315120/ and what I noticed is that BlueSpiceExtensions bring in Foundation via composer:

    "require": {
        "hallowelt/blue-spice-foundation": "~2.27"
    },

And we have composer based jobs for it, so this CI change should not be necessary.  But now that it is merged, maybe we can switch to non composer versions of the job.

Patch-set: 2