Update patch set 2
Patch Set 2: Code-Review+1

It would be nice to add that to the commit message.

In https://integration.wikimedia.org/ci/job/apps-android-wikipedia-test/4521/console I already see timestamps and have not noticed any escape sequences. It would be great to highlight errors and warnings in different colors. If this does it then I'm all for it. Is there another project already using those options, to get a bit of a preview for the ones who are curious?

Patch-set: 2
Label: Code-Review=+1