Remove UploadWizard-api-commons.wikimedia.beta.wmflabs.org
That was a smoke test for production, however nobody wanted it to run on
production commons and it has been reviewed long ago.

Whenever the job breaks on beta, nobody notices but me and I am tired a
fixing it for the sake of it.

Remove everything related to the job.

Change-Id: I4450c27b696598dff0ec60a586ad435f1fbd193b