Update patch set 2
Patch Set 2:

(1 comment)

Thanks Sam!

CI is a scary beast, it would need your repository to be ready to run and pass the commands or else Jenkins will complain on every changes.  See inline for a few hints :]

Feel free to add me as a reviewer to your changes and I will be happy to help with the preliminary setup and then enable CI reusing this change.

Patch-set: 2
Label: Code-Review=0