Update patch set 3
Patch Set 3:

Yup the CI layout looks fine :]

> I've re-added composer.json (but not composer.lock, I'm assuming? Because the composer file will only be included in the top-level site composer.local.json won't it?).

composer.lock usually we dont need it in extension. One would run composer from the root of the MediaWiki installation, it will load the composer files from all extensions and merge them then install all of that in the MediaWiki /vendor dir.  Details are on https://www.mediawiki.org/wiki/Composer  though I am not so much familiar with it.

Next step: get a change in Genealogy extension that would be able to pass the test and then we can enable the CI config :]

Patch-set: 3
Label: Verified=0