dib: apache::mod::php5 solely for Jessie
That breaks provisionning of the Trusty image due to apache2 status
trying to write to /run/lock/ which does not exist in a chroot. Somehow
Jessie is not affected.

We only have a few jobs running on Trusty and none needs Apache2, hence
move apache::mod::php5 included under a Jessie conditional.

Change-Id: Ide191bcb8ddeb11be7507f732d29428679253dac