Replace visualeditor-jsduck-jessie with npm-run-doc-jessie template
Projects should either run the generic 'jsduck' job, or if using npm,
run 'npm install' with 'npm run doc' (which puts the main command, e.g.
jsduck) in the package.json file alongside the other dependencies that
the project wants.

This avoids hardcoding a relationship on one side in the repo
and on the other side in Jenkins.

This also allows one to specify predoc/postdoc tasks in package.json
which many projects already use.

This matches oojs-ui, mediawiki-core, parsoid and others.

Change-Id: Idd396f1acaec78fe49379e688bab2b759164e90b