Replace visualeditor-jsduck-jessie with npm-run-doc-jessie template
Follows-up a11c135fb3.

Projects should either run the generic 'jsduck' job (if no dependencies
are needed and no custom or additional commands), or use npm as 'npm install'
with 'npm run doc'. This puts the main command, e.g. 'jsduck' in the
package.json file alongside the other dependencies that the project needs.

This avoids hardcoding a relationship on one side in the repo
and on the other side in Jenkins.

This also allows one to specify predoc/postdoc tasks in package.json
which many projects use already.

This matches oojs-ui, mediawiki-core, parsoid and others.

Change-Id: Idd396f1acaec78fe49379e688bab2b759164e90b