Update patch set 13
Patch Set 13:

The list of dependencies are passed as $EXT_DEPENDENCIES it is then dumped in deps.txt for zuul-cloner.  That part will work.

deps.txt is then copied to src/extensions_load.txt which is then used by the integration/jenkins.git script  mediawiki/conf.d/50_mw_ext_loader.php which has some heuristic to load the extensions.   The logic will skip the skins as not recognized, so it is going to work but it is hacky.    The skins will be added by the MediaWiki installer though given the version of MediaWiki auto discover the skins and add them to LocalSettings.php.  Not sure REL1_23 supports that though :(

Patch-set: 13